Create Nandhini_Vehicle_speed.cpp
Nandhini is tasked with creating a program to calculate and display the speed of a car based on the distance covered and the time taken.


Write a program with two classes: Vehicle as the base class and Car as the derived class, which inherits the properties from the Vehicle class for calculating the speed. The base class fetches the input as a float value, whereas the derived class calculates and prints the output as a float value.


Note: Use public inheritance


Note: This kind of question will help in clearing Wipro recruitment.
Input format :

The input consists of two floating-point numbers separated by a space. The first number represents the distance traveled by the car, and the second number represents the time taken to cover that distance, separated by a space.
Output format :

The output displays a single line of output, showing the speed of the car in kilometers per hour.


Refer to the sample output for the formatting specifications.
Code constraints :

The input distance and time should be non-negative floating-point numbers.

The input time should not be zero (to avoid division by zero).
Sample test cases :
Input 1 :

15.0 5.0

Output 1 :

Speed of car: 3 km/hr

Input 2 :

100.23 96.0

Output 2 :

Speed of car: 1.04406 km/hr