Create Amla_local_libarary.cpp
Amla is the librarian at a local library. She needs a tool to manage the library's catalog efficiently. 



She decides to use a program that can organize the book IDs in ascending order to help her locate books more easily.



Write a program that allows Sarah to input the book IDs one by one. The program should then construct a binary search tree with the book IDs and display the book IDs in ascending order using an in-order traversal.

Input format :
The first line of input consists of an integer N, the number of books in the library.

The next line consists of N positive integers separated by space, representing the unique book IDs.

Output format :
The output prints the in-order traversal of the book IDs.

Code constraints :
1 ≤ book ID ≤ 105

Sample test cases :
Input 1 :
7
53 36 74 23 41 68 87
Output 1 :
23 36 41 53 68 74 87 
Input 2 :
4
145 632 541 987
Output 2 :
145 541 632 987