Create Rahul_tree_manipulation.cpp
Rahul is working on a program for a tree manipulation application. Your task is to help Rahul implement a program that converts a given binary tree into its mirror image using preorder traversal. 



Your program should traverse the binary tree in a preordered manner and swap the left and right subtrees of each node to achieve the mirror image. 



Write a function that performs the conversion to the mirror image using a preorder traversal approach. The function should take the root of the binary tree as input and modify the tree to obtain its mirror image.

Input format :
The first line of input consists of an integer N, representing the number of nodes in the binary tree.

The second line consists of N space-separated integers representing the values of the nodes.

Output format :
The first line of output should display the preorder traversal of the original binary tree.

The second line should display the preorder traversal of the binary tree after converting it into its mirror image.

Code constraints :
1 <= node values <= 200

Sample test cases :
Input 1 :
5
4 2 6 1 3
Output 1 :
Original tree: 4 2 1 3 6 
Mirror Image: 4 6 2 3 1 
Input 2 :
8
41 52 36 51 24 64 12 25
Output 2 :
Original tree: 41 36 24 12 25 52 51 64 
Mirror Image: 41 52 64 51 36 24 25 12