Create Bindu_in-order_traversal.cpp
Bindu is working on a problem involving tree traversal, particularly the in-order traversal of a tree. She needs to construct a binary search tree (BST) from a set of integers and then traverse it in an in-order manner. 



Can you help her?



Write a program that takes a set of integers as input and constructs a binary search tree (BST) from those integers. After constructing the BST, perform an in-order traversal and display the elements in sorted order.

Input format :
The first line of input consists of an integer N, representing the number of integers to be inserted into the BST.

The second line consists of N space-separated integers, which are the elements to be inserted into the BST.

Output format :
The output displays the elements of the BST after performing an in-order traversal.

Sample test cases :
Input 1 :
5
4 2 6 1 3
Output 1 :
1 2 3 4 6 
Input 2 :
10
8 4 12 2 6 10 14 1 3 5
Output 2 :
1 2 3 4 5 6 8 10 12 14