Create Menga_ post-order_traversal.cpp
Megna is implementing a program to construct a binary tree and perform a post-order traversal on it. The binary tree will be built according to a sequence of integers provided as input. Each integer represents a node in the binary tree.


Your task is to help her construct the tree and then perform a post-order traversal to print the values of the nodes in the specified order.
Input format :

The first line of input consists of an integer, n, representing the number of integers in the sequence.

The second line of input consists of n space-separated integers, where each integer represents a node's value to be inserted into the binary tree.
Output format :

The output prints the values of the nodes in the binary tree in post-order traversal, separated by spaces.


Refer to the sample output for formatting specifications.
Code constraints :

1 <= n <= 20

1 <= elements <= 1000
Sample test cases :
Input 1 :

5
1 3 4 5 2

Output 1 :

5 2 3 4 1 

Input 2 :

5
1 7 9 5 6

Output 2 :

5 6 7 9 1