Create Dinesh.cpp
Dinesh is working in a supermarket and he is developing a program to calculate the cost of different types of items. Help him write the program that does the following:


a) Create a base class, ItemType, with one virtual function double calculateAmount()

b) Create a class called wooden that extends ItemType class with a number of items and cost as its private attributes. Obtain the data members and override the virtual function and calculate the total amount.

c) Create a class called electronics that extends ItemType class with cost as its private attribute. Obtain the data member and override the virtual function and calculate the amount with 20% discount.


Note: This question helps in clearing Infosys tests.
Input format :

The first line consists of an integer choice (1 or 2) representing the choice of item type.

If the choice is 1 (wooden items), the next line consists of two space-separated integers: noOfItems and cost, representing the number of wooden items and their individual cost, respectively.

If the choice is 2 (electronics), the next line consists of a single floating-point number cost, representing the cost of the electronic item.
Output format :

The output prints a floating-point number representing the calculated total cost of the chosen item type rounded off to two decimal places.
Code constraints :

10 < cost < 106

0 < noOfItems < 20
Sample test cases :
Input 1 :

1
5 840.5

Output 1 :

4202.50

Input 2 :

2
1800.56

Output 2 :

1440.45