Create Dancy.cpp
Dancy is developing a program for a number crunching tool that requires finding all pairs of elements whose sum is equal to a target sum. The program should take the following inputs from the user: the size of the array, the elements of the array, and the target sum.


The program should then dynamically allocate memory to store the array of integers using the new operator. Your task is to implement a function to find all pairs of elements whose sum is equal to the target sum.


Note: This kind of question will help in clearing Amcat recruitment.
Input format :

The first line represents the size of the array.

The second line represents the elements of an array separated by a space.

The third line represents the target sum.
Output format :

The output displays the pairs whose sum is equal to the target sum.

Each pair should be printed in the format: "Pair found: (element 1, element 2)".

If there is no pair, then the output displays "No pair".


Refer to the sample output for formatting specifications.
Code constraints :

1 <= size <= 100
Sample test cases :
Input 1 :

10
1 2 3 4 5 6 7 8 9 10
10

Output 1 :

Pairs with the sum 10 are:
Pair found: (6, 4)
Pair found: (7, 3)
Pair found: (8, 2)
Pair found: (9, 1)

Input 2 :

-2

Output 2 :

Size of the array should be positive!

Input 3 :

10
1 2 3 4 5 6 7 8 9 10
100

Output 3 :

Pairs with the sum 100 are:
No pair