Create Jacob.cpp
Jacob is developing a program to manage tasks in a to-do list application. He wants to implement a function to insert a new task's priority into a min heap that represents the tasks' priorities.



Write a code to help him insert new task priorities into a min heap.

Input format :
The input consists of a series of space-separated integers, each representing the priority of a new task to be inserted into the min heap.

Output format :
The output displays the priorities of the tasks after each insertion into the min heap. The priorities are separated by spaces.

Code constraints :
The maximum number of tasks that can be inserted is 100, as the min heap has a maximum capacity of 100.

Sample test cases :
Input 1 :
5 3 8 2 1
Output 1 :
1 2 8 5 3 
Input 2 :
10 7 4 6 9
Output 2 :
4 6 7 10 9