Create map.cpp
You and your friend are planning a trip and want to create a map of the places you'd like to visit. You've recorded the travel times between these places as integers, representing the hours it takes to travel. To determine which two places have the longest travel time between them, you decide to automate the process.



To do this, you need to take an array of input values, but you want to convert it into a max-heap representation. This conversion will help you efficiently find the maximum travel time among the given set of places.



Note: This kind of question will be helpful in clearing Accenture recruitment.

Input format :
The first line of the input consists of an integer n, representing the total number of recorded times for traveling between two places.

The second line of the input consists of n space-separated integer values, representing the time taken during travel from one place to another.

Output format :
The output prints the max-heap representation of the recorded time, separated by a space.

Code constraints :
1<=number of recorded time<=30

1<=value of recorded time<=50

Sample test cases :
Input 1 :
6
10 20 50 30 80 60 
Output 1 :
80 30 60 10 20 50 
Input 2 :
10
4 6 9 7 8 20 10 12 19 10
Output 2 :
20 19 10 12 10 9 4 6 7 8