Create Ethan.cpp
Ethan is working on an employee assignment management system for a company. Employees are assigned tasks based on their expertise and workload. He needs to implement a program that allows for the assignment of tasks to employees while ensuring that the employees with the highest expertise are assigned tasks first.



Write the function to help him insert a new task (an integer representing the complexity of task) into a max heap.

Input format :
The input consists of a series of space-separated integers, representing the complexity of tasks, and inserts them into the max heap.

Output format :
The output prints the max heap after each task insertion, showing the current state of the heap.

Code constraints :
The maximum heap size is 100 (defined by maxHeapSize).

Sample test cases :
Input 1 :
8 3 6 10 5 2
Output 1 :
10 8 6 3 5 2 
Input 2 :
50 30 70 20 40
Output 2 :
70 40 50 20 30