Create binary min-heap.cpp
In your digital music library, songs are organized in a binary min-heap based on their track lengths. Write a program to delete the shortest song based on its length from the heap when you're creating a playlist. After printing, rearrange the heap and print the updated list.

Input format :
The first line consists of an integer n, representing the number of songs in the playlist.

The second line consists of n space-separated integers representing the track lengths of the songs in the playlist.

Output format :
The first line displays the shortest song in the playlist after removing it.

The second line displays the elements in the updated heap separated by spaces.

Code constraints :
In this scenario, the test cases fall under the following constraints:

1 ≤ n ≤ 100

Sample test cases :
Input 1 :
5
10 15 30 20 40
Output 1 :
15 20 30 40 
Input 2 :
7
5 10 20 15 25 35 30
Output 2 :
10 15 20 30 25 35