Create Raju.cpp
Raju, a computer analyst, is working on a project that involves examining how different points in a network are connected. He needs to figure out if there's a valid way to get from one point to another based on the network setup. To solve this problem, Raju decided to create a program.



His job is to develop a program that takes a network described as a set of connections between nodes and checks if there's a way to go from a starting point to an ending point in the network. For this task, he's chosen to use the Depth-First Search (DFS) algorithm.

Input format :
The first line of input consists of two integers, n representing the number of nodes in the network and m representing the number of edges, separated by a space.

The next m lines each contain two integers, u and v, indicating an edge between nodes u and v in the network.

The last two lines of input consist of two integers, start and end, representing the starting and ending nodes for which Raju wants to find a valid path.

Output format :
The output displays one of the following two lines:



"There is a path from [start] to [end]" if a valid path exists from the starting node to the ending node.

"There is no path from [start] to [end]" if no valid path exists.



Refer to the sample output for the formatting specifications.

Code constraints :
The test cases fall under the following constraints:

1 ≤ n ≤ 10, 0 ≤ m ≤ n*(n-1)/2

0 ≤ u, v ≤ n, u ≠ v

0 ≤ start, end ≤ n, start ≠ end

Sample test cases :
Input 1 :
5 6
0 1
1 2
2 3
3 4
0 3
2 4
0
4
Output 1 :
There is a path from 0 to 4
Input 2 :
4 4
0 1
1 2
2 3
3 0
0
4
Output 2 :
There is no path from 0 to 4
Input 3 :
6 7
0 1
0 2
1 3
2 3
3 4
4 5
5 2
0
5
Output 3 :
There is a path from 0 to 5