Create Rithu.cpp
Rithu is currently working on a system to monitor and analyze dependencies between tasks within a project. Each task is represented as a node in a directed graph, and dependencies between tasks are represented as directed edges between nodes.



The goal is to help Rithu find the longest chain of tasks linked together in the project using the Depth-First Search (DFS) approach.

Input format :
The first line of input consists of two integers, "n" and "m," separated by a space. "n" represents the number of tasks in the project, and "m" represents the number of task dependencies (directed edges).

The next "m" lines each contain two integers, "u" and "v," separated by a space, indicating a task dependency where task "u" depends on task "v."

Output format :
The output displays a single integer representing the length of the longest chain of tasks in the project.



Refer to the sample output for the formatting specifications.

Code constraints :
The test cases will fall under the following constraints:

2 ≤ n ≤ 10

0 ≤ m ≤ n*(n-1)/2

1 ≤ u, v ≤ n, u ≠ v

Sample test cases :
Input 1 :
4 5
2 3 
1 3
1 4
1 2
3 4
Output 1 :
3
Input 2 :
3 2
0 1
1 2
Output 2 :
1