Create Truck_index.cpp
Imagine you own a logistics company where you run N number of trucks. Each truck has a unique loading capacity. Now use quadratic probing to arrange the trucks in the truck shed.



Assume the hashtable size is 10. Write a code to find out in which index the truck that has the highest loading capacity stands.



Implement a hash function that calculates the hash index for each key using the modulo operation with the table size (10). The hash index should be calculated as follows: hashIndex = key % tableSize.

Input format :
The first line of input consists of an integer N, representing the number of trucks.

The second line consists of N integers, representing the loading capacity of each truck.

Output format :
The output prints the index of the truck that has the highest loading capacity after hashing.



Refer to the sample output for the exact text and format.

Code constraints :
table size = 10

Sample test cases :
Input 1 :
5
5 10 15 20 25
Output 1 :
The lorry that has high load capacity is in index 9.
Input 2 :
8
14 24 25 37 45 57 67 51
Output 2 :
The lorry that has high load capacity is in index 2.